Added a GridLoader to load grids values for new games

+++ src/main/java/ar/fiuba/tdd/tp/cell/NumericalCell.java
@@ -18,7 +18,7 @@ public class NumericalCell extends Cell {
     }
 
     public int getValue() {
-        return this.value;
+        return this.value.intValue();
     }
 
     public Boolean isEmpty() {
